split up daemon api and state
