+++ pkg/daemon/api/readers.go
@@ -1,0 +1,29 @@
+package api
+
+import (
+	"encoding/json"
+	"net/http"
+
+	"github.com/rs/zerolog/log"
+	"github.com/wizzomafizzo/tapto/pkg/daemon/state"
+)
+
+type ReaderWriteRequest struct {
+	Text string `json:"text"`
+}
+
+func handleReaderWrite(st *state.State) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		log.Info().Msg("received reader write request")
+
+		var req ReaderWriteRequest
+		err := json.NewDecoder(r.Body).Decode(&req)
+		if err != nil {
+			log.Error().Err(err).Msg("error decoding request")
+			http.Error(w, err.Error(), http.StatusBadRequest)
+			return
+		}
+
+		st.SetWriteRequest(req.Text)
+	}
+}
