+++ pkg/daemon/state/state.go
@@ -1,4 +1,4 @@
-package daemon
+package state
 
 import (
 	"os"
